Use a mask to understand which features are categorical
I wanted to import manually the datasets of Open CC-18 through their csv because if you access the data through the openml API the categorical features are already encoded with an OrdinalEncoder. Unfortunatly, I realized that in the csv files the target column is not always named as "class" and there are a lot column unusable like "id", "timestamp", etc. because they could affect the result. We wanted to let the SMBO algorithm decide the better encoding method (OneHotEncoding or OrdinalEncoding) but for this experiment we just allow the algorithm to decide if apply the OneHotEncoding to the categorical features already encoded with the OrdinalEncoding or not.